feat(storybook): search page refactor, allow overrides + new story
