+++ client/src/app/pages/search/components/SearchTabs.tsx
@@ -1,0 +1,172 @@
+import React from "react";
+
+import {
+  Badge,
+  Card,
+  CardBody,
+  Popover,
+  Split,
+  SplitItem,
+  Tab,
+  TabAction,
+  Tabs,
+  TabTitleText,
+} from "@patternfly/react-core";
+
+import { PackageTable } from "@app/pages/package-list/package-table";
+import { AdvisoryTable } from "@app/pages/advisory-list/advisory-table";
+import { SbomTable } from "@app/pages/sbom-list/sbom-table";
+import { VulnerabilityTable } from "@app/pages/vulnerability-list/vulnerability-table";
+import HelpIcon from "@patternfly/react-icons/dist/esm/icons/help-icon";
+import { FilterPanel } from "@app/components/FilterPanel";
+
+interface SearchTabsProps {
+  filterPanelProps: {
+    advisoryFilterPanelProps?: any;
+    packageFilterPanelProps?: any;
+    sbomFilterPanelProps?: any;
+    vulnerabilityFilterPanelProps?: any;
+  };
+  packageTotalCount: number;
+  sbomTotalCount: number;
+  vulnerabilityTotalCount: number;
+}
+
+export const SearchTabs: React.FC<SearchTabsProps> = ({
+  filterPanelProps,
+  packageTotalCount,
+  sbomTotalCount,
+  vulnerabilityTotalCount,
+}) => {
+  const [activeTabKey, setActiveTabKey] = React.useState<string | number>(0);
+  const {
+    advisoryFilterPanelProps,
+    packageFilterPanelProps,
+    sbomFilterPanelProps,
+    vulnerabilityFilterPanelProps,
+  } = filterPanelProps;
+
+  const handleTabClick = (
+    _event: React.MouseEvent<any> | React.KeyboardEvent | MouseEvent,
+    tabIndex: string | number
+  ) => {
+    setActiveTabKey(tabIndex);
+  };
+
+  const sbomPopoverRef = React.createRef<HTMLElement>();
+
+  const sbomPopover = (popoverRef: React.RefObject<any>) => (
+    <Popover
+      bodyContent={
+        <div>Software Bill of Materials for Products and Containers.</div>
+      }
+      position={"right"}
+      triggerRef={popoverRef}
+    />
+  );
+
+  return (
+    <Split hasGutter>
+      <SplitItem>
+        <Card isFullHeight>
+          <CardBody style={{ width: 241 }}>
+            {activeTabKey === 0 ? (
+              <FilterPanel
+                omitFilterCategoryKeys={[""]}
+                {...sbomFilterPanelProps}
+              />
+            ) : activeTabKey === 1 ? (
+              <FilterPanel
+                omitFilterCategoryKeys={[""]}
+                {...packageFilterPanelProps}
+              />
+            ) : activeTabKey === 2 ? (
+              <FilterPanel
+                omitFilterCategoryKeys={[""]}
+                {...vulnerabilityFilterPanelProps}
+              />
+            ) : activeTabKey === 3 ? (
+              <FilterPanel
+                omitFilterCategoryKeys={[""]}
+                {...advisoryFilterPanelProps}
+              />
+            ) : null}
+          </CardBody>
+        </Card>
+      </SplitItem>
+      <SplitItem isFilled>
+        <Tabs
+          isBox
+          activeKey={activeTabKey}
+          onSelect={handleTabClick}
+          aria-label="Tabs"
+          role="region"
+        >
+          <Tab
+            eventKey={0}
+            title={
+              <TabTitleText>
+                SBOMs{"  "}
+                <Badge screenReaderText="Search Result Count">
+                  {sbomTotalCount}
+                </Badge>
+              </TabTitleText>
+            }
+            actions={
+              <>
+                <TabAction
+                  aria-label={`SBOM help popover`}
+                  ref={sbomPopoverRef}
+                >
+                  <HelpIcon />
+                </TabAction>
+                {sbomPopover(sbomPopoverRef)}
+              </>
+            }
+          >
+            <SbomTable />
+          </Tab>
+          <Tab
+            eventKey={1}
+            title={
+              <TabTitleText>
+                Packages{"  "}
+                <Badge screenReaderText="Search Result Count">
+                  {packageTotalCount}
+                </Badge>
+              </TabTitleText>
+            }
+          >
+            <PackageTable />
+          </Tab>
+          <Tab
+            eventKey={2}
+            title={
+              <TabTitleText>
+                CVEs{"  "}
+                <Badge screenReaderText="Search Result Count">
+                  {vulnerabilityTotalCount}
+                </Badge>
+              </TabTitleText>
+            }
+          >
+            <VulnerabilityTable />
+          </Tab>
+          <Tab
+            eventKey={3}
+            title={
+              <TabTitleText>
+                Advisories{"  "}
+                <Badge screenReaderText="Advisory Result Count">
+                  {vulnerabilityTotalCount}
+                </Badge>
+              </TabTitleText>
+            }
+          >
+            <AdvisoryTable />
+          </Tab>
+        </Tabs>
+      </SplitItem>
+    </Split>
+  );
+};
