+++ client/src/mocks/data/advisory/details/urn%3Auuid%3A03bb16dc-3cff-4a7d-8393-9a6a7124ecc2.json
@@ -1,0 +1,41 @@
+{
+  "uuid": "urn:uuid:03bb16dc-3cff-4a7d-8393-9a6a7124ecc2",
+  "identifier": "CVE-2023-1664",
+  "issuer": {
+    "id": "4aa82c3b-fc6e-41d7-ba55-e1c37853831a",
+    "name": "redhat",
+    "cpe_key": null,
+    "website": null
+  },
+  "published": "2023-05-26T00:00:00Z",
+  "modified": "2024-08-02T05:57:24.969Z",
+  "withdrawn": null,
+  "title": "A flaw was found in Keycloak. This flaw depends on a non-default configuration \"Revalidate Client Certificate\" to be enabled and the reverse proxy is not validating the certificate before Keycloak. Using this method an attacker may choose the certificate which will be validated by the server. If this happens and the KC_SPI_TRUSTSTORE_FILE_FILE variable is missing/misconfigured, any trustfile may be accepted with the logging information of \"Cannot validate client certificate trust: Truststore not available\". This may not impact availability as the attacker would have no access to the server, but consumer applications Integrity or Confidentiality may be impacted considering a possible access to them. Considering the environment is correctly set to use \"Revalidate Client Certificate\" this flaw is avoidable.",
+  "labels": {
+    "type": "cve"
+  },
+  "sha256": "sha256:72b745cfcceed4e905de9d15f6826bf70fd5429fcb5bb31147522f7b7e6389cd",
+  "sha384": "sha384:f559a4d9e893291384e6acaae203d590c979786311cc19351e89e1467baf1d95870f7783d189922e955b15683e36fc48",
+  "sha512": "sha512:b536afb9cde0e25a3294c7cb3d72371453dc42dca2a732b34fafe7ce8a75a893a4a4822908eadaad9e75e182f1a4b0b31a92a0b104df3485d0541d050e8fc2ab",
+  "size": 3162,
+  "vulnerabilities": [
+    {
+      "normative": true,
+      "identifier": "CVE-2023-1664",
+      "title": "A flaw was found in Keycloak. This flaw depends on a non-default configuration \"Revalidate Client Certificate\" to be enabled and the reverse proxy is not validating the certificate before Keycloak. Using this method an attacker may choose the certificate which will be validated by the server. If this happens and the KC_SPI_TRUSTSTORE_FILE_FILE variable is missing/misconfigured, any trustfile may be accepted with the logging information of \"Cannot validate client certificate trust: Truststore not available\". This may not impact availability as the attacker would have no access to the server, but consumer applications Integrity or Confidentiality may be impacted considering a possible access to them. Considering the environment is correctly set to use \"Revalidate Client Certificate\" this flaw is avoidable.",
+      "description": "A flaw was found in Keycloak. This flaw depends on a non-default configuration \"Revalidate Client Certificate\" to be enabled and the reverse proxy is not validating the certificate before Keycloak. Using this method an attacker may choose the certificate which will be validated by the server. If this happens and the KC_SPI_TRUSTSTORE_FILE_FILE variable is missing/misconfigured, any trustfile may be accepted with the logging information of \"Cannot validate client certificate trust: Truststore not available\". This may not impact availability as the attacker would have no access to the server, but consumer applications Integrity or Confidentiality may be impacted considering a possible access to them. Considering the environment is correctly set to use \"Revalidate Client Certificate\" this flaw is avoidable.",
+      "reserved": "2023-03-27T00:00:00Z",
+      "published": "2023-05-26T00:00:00Z",
+      "modified": "2024-08-02T05:57:24.969Z",
+      "withdrawn": null,
+      "discovered": null,
+      "released": null,
+      "cwes": ["CWE-295"],
+      "severity": "none",
+      "score": 0.0,
+      "cvss3_scores": []
+    }
+  ],
+  "average_severity": null,
+  "average_score": null
+}
