+++ client/src/mocks/data/advisory/details/urn%3Auuid%3A459c504b-7e09-4ea9-9cbb-baa8ce040e83.json
@@ -1,0 +1,41 @@
+{
+  "uuid": "urn:uuid:459c504b-7e09-4ea9-9cbb-baa8ce040e83",
+  "identifier": "CVE-2023-1436",
+  "issuer": {
+    "id": "bb1dc009-4428-4163-8dee-ddac5823f1fb",
+    "name": "JFROG",
+    "cpe_key": null,
+    "website": null
+  },
+  "published": "2023-03-16T20:59:51.072Z",
+  "modified": "2024-08-02T05:49:11.438Z",
+  "withdrawn": null,
+  "title": "Infinite recursion in Jettison leads to denial of service when creating a crafted JSONArray",
+  "labels": {
+    "type": "cve"
+  },
+  "sha256": "sha256:dfe61f4eabad4795e6bbc9074fc0f51024275809128410cd803fc8d5cfde89b4",
+  "sha384": "sha384:d35fd9015b35864544b11e53bf8cc923da6facd641fed4d99152e06f0fabc83e2b4eb7f2b062597d91fbbe98dd177b67",
+  "sha512": "sha512:b905135296652ff7b27620150fe12e67908ad7b0b5299ba9974735f7fc2608ebf2a7188c30570f627a3c0646cd614b54c727a3d7b37e90b0e122a1ff5c81fffc",
+  "size": 4531,
+  "vulnerabilities": [
+    {
+      "normative": true,
+      "identifier": "CVE-2023-1436",
+      "title": "Infinite recursion in Jettison leads to denial of service when creating a crafted JSONArray",
+      "description": "An infinite recursion is triggered in Jettison when constructing a JSONArray from a Collection that contains a self-reference in one of its elements. This leads to a StackOverflowError exception being thrown.\n\n",
+      "reserved": "2023-03-16T20:44:44.527Z",
+      "published": "2023-03-16T20:59:51.072Z",
+      "modified": "2024-08-02T05:49:11.438Z",
+      "withdrawn": null,
+      "discovered": null,
+      "released": null,
+      "cwes": ["CWE-674"],
+      "severity": "none",
+      "score": 0.0,
+      "cvss3_scores": []
+    }
+  ],
+  "average_severity": null,
+  "average_score": null
+}
