+++ client/src/mocks/data/advisory/details/urn%3Auuid%3A671dd85b-409f-4509-9a50-c4b2404ac10a.json
@@ -1,0 +1,41 @@
+{
+  "uuid": "urn:uuid:671dd85b-409f-4509-9a50-c4b2404ac10a",
+  "identifier": "CVE-2023-0481",
+  "issuer": {
+    "id": "8e0a1182-9579-4ab2-8913-9a1c37b224e3",
+    "name": "redhat",
+    "cpe_key": null,
+    "website": null
+  },
+  "published": "2023-02-24T00:00:00Z",
+  "modified": "2024-08-02T05:10:56.412Z",
+  "withdrawn": null,
+  "title": "In RestEasy Reactive implementation of Quarkus the insecure File.createTempFile() is used in the FileBodyHandler class which creates temp files with insecure permissions that could be read by a local user.",
+  "labels": {
+    "type": "cve"
+  },
+  "sha256": "sha256:9453778b1ff60cff4d9bf1455615dc5035889e225ece201781001f29255aa779",
+  "sha384": "sha384:b2d77c3560c861da161af3ab5bd20775298e39d014b5d43b0be81147c2de5ecb02844fefcebe179d54c6313278e7e0dd",
+  "sha512": "sha512:2953fb9855e78a9a59addc6a0337d0755efb5347a653556e1f7f08eb260814bc7104bb26f89dc16ffeef56a4f1084898b98c6c0864db5f6d924cde3acdf6fd18",
+  "size": 2530,
+  "vulnerabilities": [
+    {
+      "normative": true,
+      "identifier": "CVE-2023-0481",
+      "title": "In RestEasy Reactive implementation of Quarkus the insecure File.createTempFile() is used in the FileBodyHandler class which creates temp files with insecure permissions that could be read by a local user.",
+      "description": "In RestEasy Reactive implementation of Quarkus the insecure File.createTempFile() is used in the FileBodyHandler class which creates temp files with insecure permissions that could be read by a local user.",
+      "reserved": "2023-01-24T00:00:00Z",
+      "published": "2023-02-24T00:00:00Z",
+      "modified": "2024-08-02T05:10:56.412Z",
+      "withdrawn": null,
+      "discovered": null,
+      "released": null,
+      "cwes": ["CWE-378"],
+      "severity": "none",
+      "score": 0.0,
+      "cvss3_scores": []
+    }
+  ],
+  "average_severity": null,
+  "average_score": null
+}
