+++ client/src/mocks/data/advisory/details/urn%3Auuid%3A673acfc8-ea7d-4c6d-aff9-20cf70caade0.json
@@ -1,0 +1,41 @@
+{
+  "uuid": "urn:uuid:673acfc8-ea7d-4c6d-aff9-20cf70caade0",
+  "identifier": "CVE-2023-0482",
+  "issuer": {
+    "id": "4aa82c3b-fc6e-41d7-ba55-e1c37853831a",
+    "name": "redhat",
+    "cpe_key": null,
+    "website": null
+  },
+  "published": "2023-02-17T00:00:00Z",
+  "modified": "2024-08-02T05:10:56.348Z",
+  "withdrawn": null,
+  "title": "In RESTEasy the insecure File.createTempFile() is used in the DataSourceProvider, FileProvider and Mime4JWorkaround classes which creates temp files with insecure permissions that could be read by a local user.",
+  "labels": {
+    "type": "cve"
+  },
+  "sha256": "sha256:0f6262e341bdebc54c48d65342bcbfd8c941b661c0291d4fde3cf724fdda741a",
+  "sha384": "sha384:8cc216134f63b18debfb447d0669b786db108b7cc8b1d19b262d32c48d9e0d306cbbf884dad0e1acf0983b71640569f4",
+  "sha512": "sha512:9a56d4e4de0225274022953ba8f4df1e3eb16cfd3bee96ee459b14398d8b075247427fa1882771cbb05fb3fa86f44797a25dd7912183d9878d8de7c096a5d9b1",
+  "size": 3009,
+  "vulnerabilities": [
+    {
+      "normative": true,
+      "identifier": "CVE-2023-0482",
+      "title": "In RESTEasy the insecure File.createTempFile() is used in the DataSourceProvider, FileProvider and Mime4JWorkaround classes which creates temp files with insecure permissions that could be read by a local user.",
+      "description": "In RESTEasy the insecure File.createTempFile() is used in the DataSourceProvider, FileProvider and Mime4JWorkaround classes which creates temp files with insecure permissions that could be read by a local user.",
+      "reserved": "2023-01-24T00:00:00Z",
+      "published": "2023-02-17T00:00:00Z",
+      "modified": "2024-08-02T05:10:56.348Z",
+      "withdrawn": null,
+      "discovered": null,
+      "released": null,
+      "cwes": ["CWE-378"],
+      "severity": "none",
+      "score": 0.0,
+      "cvss3_scores": []
+    }
+  ],
+  "average_severity": null,
+  "average_score": null
+}
