+++ client/src/mocks/data/purl/details/b9a43108-525d-59ea-bc31-ff217d4c7925.json
@@ -1,0 +1,54 @@
+{
+  "uuid": "b9a43108-525d-59ea-bc31-ff217d4c7925",
+  "purl": "pkg:oci/rhacs-roxctl-rhel8@sha256:ad2b63ce031af12d8c31eb40652e324ea9cc5891624c4e1683c00d2948825c0f?arch=amd64&repository_url=registry.redhat.io/advanced-cluster-security/rhacs-roxctl-rhel8&tag=3.74.7-3",
+  "version": {
+    "uuid": "001776b0-c8f7-5bdd-ad82-936ff42dc1d8",
+    "purl": "pkg:oci/rhacs-roxctl-rhel8@sha256:ad2b63ce031af12d8c31eb40652e324ea9cc5891624c4e1683c00d2948825c0f",
+    "version": "sha256:ad2b63ce031af12d8c31eb40652e324ea9cc5891624c4e1683c00d2948825c0f"
+  },
+  "base": {
+    "uuid": "39082781-f0ba-56c3-86c5-4e7c843bbc94",
+    "purl": "pkg:oci/rhacs-roxctl-rhel8"
+  },
+  "advisories": [
+    {
+      "uuid": "urn:uuid:5b98f727-1145-4eca-8176-d3f2a32e0fce",
+      "identifier": "https://www.redhat.com/#CVE-2023-44487",
+      "issuer": {
+        "id": "5e9c87c9-2e6b-453b-8b27-6c80dfc7d7ff",
+        "name": "Red Hat Product Security",
+        "cpe_key": null,
+        "website": null
+      },
+      "published": "2023-10-10T00:00:00Z",
+      "modified": "2023-11-21T11:59:10Z",
+      "withdrawn": null,
+      "title": "HTTP/2: Multiple HTTP/2 enabled web servers are vulnerable to a DDoS attack (Rapid Reset Attack)",
+      "labels": {
+        "type": "csaf"
+      },
+      "status": [
+        {
+          "vulnerability": {
+            "normative": true,
+            "identifier": "CVE-2023-44487",
+            "title": "The HTTP/2 protocol allows a denial of service (server resource consumption) because request cancellation can reset many streams quickly, as exploited in the wild in August through October 2023.",
+            "description": "The HTTP/2 protocol allows a denial of service (server resource consumption) because request cancellation can reset many streams quickly, as exploited in the wild in August through October 2023.",
+            "reserved": "2023-09-29T00:00:00Z",
+            "published": "2023-10-10T00:00:00Z",
+            "modified": "2024-08-19T07:48:04.546Z",
+            "withdrawn": null,
+            "discovered": null,
+            "released": null,
+            "cwes": []
+          },
+          "status": "not_affected",
+          "context": {
+            "cpe": "cpe:/a:redhat:advanced_cluster_security:3.74:*:el8:*"
+          }
+        }
+      ]
+    }
+  ],
+  "licenses": []
+}
