+++ client/src/stories/v1.2.x/cve-details.stories.tsx
